backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.1",
    "beautifulsoup4>=4.14.3",
    "celery>=5.6.2",
    "fastapi[all]>=0.128.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "pytz>=2025.1",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
]
